+++ youtube-autolike-script.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name         YouTube autolike
 // @namespace    https://github.com/blackstar0169/youtube-autolike-script
-// @version      1.7
+// @version      1.8
 // @description  Auto-like YouTube video if you are subscribed to the author's channel. Or can set an option to like all videos regardless if you are subrscribed or not.
 // @author       blackstar0169
 // @match        https://www.youtube.com/*
@@ -12,15 +12,15 @@
 // @license      CC-BY-NC-SA-4.0
 // @downloadURL  https://update.greasyfork.org/scripts/423125/YouTube%20autolike.user.js
 // @updateURL    https://update.greasyfork.org/scripts/423125/YouTube%20autolike.meta.js
-// @require      https://raw.github.com/odyniec/MonkeyConfig/master/monkeyconfig.js#sha256=65494c58e6b9f19e7f62683a1b352a943d58c59f6c2ecb184ac422934d48281f
+// @require      https://raw.github.com/blackstar0169/MonkeyConfig/master/monkeyconfig.js
 // ==/UserScript==
 
 (function() {
     'use strict';
 
     var watcher = null;
-    var likeSelector = '.YtLikeButtonViewModelHost button';
-    var subscribeSelector = '#subscribe-button button';
+    const likeSelector = '.ytLikeButtonViewModelHost button';
+    const subscribeSelector = '#subscribe-button button';
 
     // Config
     const config = new MonkeyConfig({
@@ -35,7 +35,7 @@
     });
 
     function getObjectProperty(obj, keyStr) {
-        var keys = keyStr.split('.');
+        const keys = keyStr.split('.');
 
         for (var i = 0; i < keys.length; i++) {
             var key = keys[i];
@@ -54,13 +54,21 @@
 
     // Check if the video is already liked
     function isLiked() {
-        var btn = document.querySelector(likeSelector);
-        return btn ? btn.getAttribute('aria-pressed') == 'true' : false;
+        const btn = document.querySelector(likeSelector);
+        if (btn === null) {
+            console.error("[YouTube autolike] can not find like button !");
+            return false;
+        }
+        return btn.getAttribute('aria-pressed') == 'true';
     }
 
     function isSubscribed() {
-        var btn = document.querySelector(subscribeSelector);
-        return (btn && btn.classList.contains('yt-spec-button-shape-next--tonal')) ||
+        const btn = document.querySelector(subscribeSelector);
+        if (btn === null) {
+            console.error("[YouTube autolike] can not find subscribe button !");
+            return false;
+        }
+        return btn.classList.contains('yt-spec-button-shape-next--tonal') ||
             getObjectProperty(window.ytInitialPlayerResponse, 'annotations.0.playerAnnotationsExpandedRenderer.featuredChannel.subscribeButton.subscribeButtonRenderer.subscribed');
     }
 
@@ -84,11 +92,9 @@
     function startWatcher() {
         if (watcher === null) {
             watcher = setInterval(function () {
-                if (document.querySelector(subscribeSelector)) {
-                    if (isLiked() || autoLike()) {
-                        clearInterval(watcher);
-                        watcher = null;
-                    }
+                if (isLiked() || autoLike()) {
+                    clearInterval(watcher);
+                    watcher = null;
                 }
             }, 1000);
         }